backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["rcidral <ricidralmfo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.25.1"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
